Add files via upload
IoT is commonly used to transmit sensing data which is then stored in the database. When IoT network is large, the more data is stored. Furthermore, it causes various problems including data security and data integrity issues. In addition, Blockchain Technology is a decentralized system which all nodes in the blockchain network stores a copy of the recorded information. To overcome this problem,it proposes the implementation of blockchain in IoT sensor data to reduce problems such as data manipulation. It shows that Ethereum-based blockchain with using consensus proof-of- work as a blockchain network is able to ensure integrity of irrigation data sensor.